requirements.txt
lightning
torch
torchaudio
timm
wandb
numpy
pandas
scikit-learn
python-dotenv
tqdm
